+++ package.json
@@ -9,7 +9,7 @@
   "description": "REST service starter for RS School",
   "main": "src/server.js",
   "scripts": {
-    "start": "nodemon src/fastify/server.js",
+    "start": "nodemon src/server.js",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
@@ -48,10 +48,10 @@
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "express": "4.17.1",
-    "fastify": "^3.24.1",
     "swagger-ui-express": "4.1.6",
-    "uuid": "8.3.2",
-    "yamljs": "0.3.0"
+    "yamljs": "0.3.0",
+    "fastify": "^3.24.1",
+    "uuid": "8.3.2"
   },
   "devDependencies": {
     "chai": "4.3.4",
