--- src/resources/users/user.service.js
@@ -1,5 +1,0 @@
-const usersRepo = require('./user.memory.repository');
-
-const getAll = () => usersRepo.getAll();
-
-module.exports = { getAll };
